[styles] Overload function signature instead of conditional (#19320)
