--- packages/material-ui-styles/test/IsEmptyInterface.spec.ts
@@ -1,42 +1,0 @@
-/**
- * unit testing for IsEmptyInterface utility type
- */
-import { IsEmptyInterface } from '@material-ui/types';
-
-// $ExpectType true
-type EmptyInterfaceIsValid = IsEmptyInterface<{}>;
-
-// $ExpectType false
-type ObjectIsValid = IsEmptyInterface<object>;
-
-// $ExpectType false
-type NullishIsValid = IsEmptyInterface<null | undefined>;
-
-// $ExpectType false
-type StringIsValid = IsEmptyInterface<string>;
-
-// $ExpectType false
-type NumberIsValid = IsEmptyInterface<number>;
-
-// $ExpectType false
-type SymbolIsValid = IsEmptyInterface<symbol>;
-
-// $ExpectType false
-type NeverIsValid = IsEmptyInterface<never>;
-
-// $ExpectType false
-type UnknownIsValid = IsEmptyInterface<unknown>;
-
-// $ExpectType false
-type OptionalIsValid = IsEmptyInterface<{x?: number}>;
-
-const noop = () => {};
-// $ExpectType false
-type FunctionIsValid = IsEmptyInterface<typeof noop>;
-
-// $ExpectType false
-type RecordIsValid = IsEmptyInterface<Record<'foo' | 'bar', number>>;
-
-const someObject = { foo: 5 };
-// $ExpectType false
-type ObjectWithPropertiesIsValid = IsEmptyInterface<typeof someObject>;
