+++ packages/material-ui/src/styles/index.d.ts
@@ -7,7 +7,7 @@ export {
 } from './createPalette';
 export { default as createStyles } from './createStyles';
 export { TypographyStyle } from './createTypography';
-export { default as makeStyles, StylesHook } from './makeStyles';
+export { default as makeStyles } from './makeStyles';
 export { default as responsiveFontSizes } from './responsiveFontSizes';
 export { ComponentsPropsList } from './props';
 export * from './transitions';
