+++ packages/material-ui/src/styles/makeStyles.d.ts
@@ -1,8 +1,18 @@
 import { Theme as DefaultTheme } from './createMuiTheme';
-import { Styles, WithStylesOptions } from '@material-ui/styles/withStyles';
-import { StylesHook } from '@material-ui/styles/makeStyles';
+import { ClassNameMap, Styles, WithStylesOptions } from '@material-ui/styles/withStyles';
+
 import { Omit } from '@material-ui/types';
 
+/**
+ * `makeStyles` where the passed `styles` do not depend on props
+ */
+export default function makeStyles<Theme = DefaultTheme, ClassKey extends string = string>(
+  style: Styles<Theme, {}, ClassKey>,
+  options?: Omit<WithStylesOptions<Theme>, 'withTheme'>,
+): (props?: any) => ClassNameMap<ClassKey>;
+/**
+ * `makeStyles` where the passed `styles` do depend on props
+ */
 export default function makeStyles<
   Theme = DefaultTheme,
   Props extends {} = {},
@@ -10,6 +20,4 @@ export default function makeStyles<
 >(
   styles: Styles<Theme, Props, ClassKey>,
   options?: Omit<WithStylesOptions<Theme>, 'withTheme'>,
-): StylesHook<Styles<Theme, Props, ClassKey>>;
-
-export { StylesHook };
+): (props: Props) => ClassNameMap<ClassKey>;
